tidy

+++ images/03.ServiceImages/auth/home/add_service.sh
@@ -37,15 +37,15 @@ if test -z $command
 
 if  test $command = "access"
 	then
+	if ! test -f /home/auth/static/access/$service/access
+	 then
 		mkdir -p /home/auth/static/access/$service/
 		cp /home/get_access.sh /home/auth/static/scripts/$service/
-		chmod u+x /home/auth/static/scripts/$service/get_access.sh 
-	
-		echo "command=\"/home/auth/static/scripts/$service/get_access.sh\",no-port-forwarding,no-X11-forwarding,no-agent-forwarding,no-pty ssh-rsa $pubkey auth" >  /home/auth/keys/${service}_${command}_authorized_keys	
-	
+		chmod u+x /home/auth/static/scripts/$service/get_access.sh 	
+		echo "command=\"/home/auth/static/scripts/$service/get_access.sh\",no-port-forwarding,no-X11-forwarding,no-agent-forwarding,no-pty ssh-rsa $pubkey auth" >  /home/auth/keys/${service}_${command}_authorized_keys		
 		password=`dd if=/dev/urandom count=6 bs=1  | od -h | awk '{ print $2$3$4}'`
 		echo "create user 'auth_$service'@'%' identified by '$password';;" | mysql -h $dbhost -u $dbuser --password=$dbpasswd $dbname 
-	echo "create user 'auth_$service'@'%' identified by '$password';;" 
+	    echo "create user 'auth_$service'@'%' identified by '$password';;" 
 		echo "GRANT SELECT on auth.* to 'auth_$service'@'%';" | mysql -h $dbhost -u $dbuser --password=$dbpasswd $dbname 
 		echo "GRANT SELECT on auth.* to 'auth_$service'@'%';"
 		#echo "create user 'auth_$service'@'%' identified by '$password';
@@ -53,9 +53,9 @@ if  test $command = "access"
 		#	GRANT SELECT on auth.* to 'auth_$service'@'%'; | mysql -h $dbhost -u $dbuser --password=$dbpasswd $dbname " >>/tmp/add_access.log
 		#echo ":db_username=auth_$service:db_password=$password:database_name=$dbname:db_host=$dbhost:" > /home/auth/static/access/$service/access
 		echo '{"db_username":"auth_'$service'","db_password":"'$password'","database_name":"'$dbname'","db_host":"'$dbhost'"}' > /home/auth/static/access/$service/access
-		
-	else
-			echo "command=\"/home/auth/static/scripts/${service}/${command}_service.sh\",no-port-forwarding,no-X11-forwarding,no-agent-forwarding,no-pty ssh-rsa $pubkey auth" >  /home/auth/keys/${service}_${command}_authorized_keys	
+	fi	
+  else
+	echo "command=\"/home/auth/static/scripts/${service}/${command}_service.sh\",no-port-forwarding,no-X11-forwarding,no-agent-forwarding,no-pty ssh-rsa $pubkey auth" >  /home/auth/keys/${service}_${command}_authorized_keys	
 	fi
 
 #
